test: 테스트 편의성을 위해 Path 관련 domain 에서 사용 할 테스트 베이스 클래스 생성 및 기본작업 이관
